Refactor sprite data and component structures for improved organization and functionality.
- Updated `human_body0.spridat.ron` to correct the category from "head" to "body" and adjusted the `rows_cols` values. Added a `frame_size` and `offset` properties to enhance sprite rendering. Modified the `offset_children` to better position child sprites relative to the parent.

- In `human_head0.spridat.ron`, added a `parent_cat` field to establish a relationship with the body category. Removed the previous `parent_cat` entry to avoid redundancy. This change helps in organizing sprite hierarchy more effectively.

- Changed `GameZindex` in `common_components.rs` from `f32` to `i32` for consistency with other components. This adjustment ensures that z-index values are handled uniformly across the application. It also simplifies the management of z-index values in the rendering pipeline.

- Updated `being_components.rs` to include `FacingDirection` in the requirements for the `Being` struct. This addition allows for better control over the direction entities are facing during gameplay. It enhances the interaction between movement and animation systems.

- Refactored `update_direction` in `being_systems.rs` to streamline the handling of direction updates. Removed unnecessary entity references to simplify the query and improve performance. This change also ensures that the facing direction is updated directly without additional checks.

- Improved error logging in `animate_sprite` within `animation_systems.rs` by adding a warning message when an animation is not found. This change enhances debugging capabilities by providing clearer feedback during runtime. It helps developers quickly identify missing animations.

- Removed the `OffsetGivenByParent` struct from `sprite_components.rs` to clean up unused code. This change reduces clutter and improves maintainability of the codebase. It also ensures that only relevant components are retained for sprite management.

- In `game_components.rs`, added a random generation method for `FacingDirection` to facilitate dynamic behavior in entities. This method allows entities to have a random initial facing direction, enhancing gameplay variability. The default implementation now utilizes this random method for initialization.

- Adjusted the z-index calculation in `game_systems.rs` to apply a scaling factor for better depth management. This change ensures that z-index values are appropriately scaled for rendering. It improves the visual layering of entities in the game world.

- Modified `main.rs` to set the image plugin to use nearest filtering for better pixel art rendering. This change enhances the visual quality of sprites when scaled. It ensures that the game maintains a consistent aesthetic throughout.

(Commit message generated with GitHub Copilot)

+++ src/common/common_components.rs
@@ -8,4 +8,4 @@ use serde::{Deserialize, Serialize};
 
 
 #[derive(Component, Debug,)]
-pub struct GameZindex(pub f32);
+pub struct GameZindex(pub i32);
